create auth, validate, update with image or not

+++ src/controllers/news.js
@@ -6,72 +6,79 @@ module.exports = {
   createNews: async (req, res) => {
     try {
       // const image = req.file
-      console.log(req.file)
-      const schema = joi.object({
-        title: joi.string().required(),
-        news: joi.string().required(),
-        category: joi.string().required()
-      })
-      const { value, error } = schema.validate(req.body)
-      const { title, news, category } = value
-      if (error) {
-        res.send({
-          success: false,
-          message: `${error}`
+      const { role } = req.user.jwtToken
+      if (role === 'admin' || role === 'jurnalis') {
+        const schema = joi.object({
+          title: joi.string().required(),
+          news: joi.string().required(),
+          category: joi.string().required()
         })
-      } else {
-        if (req.file === undefined) {
-          const checkNews = await News.findAll({ where: { title: title } })
-          if (checkNews.length > 0) {
-            res.send({
-              success: false,
-              message: 'News alreadey exists'
-            })
-          } else {
-            const data = {
-              title, news, category
-            }
-            const results = await News.create(data)
-            if (results) {
-              res.send({
-                success: true,
-                message: 'Success add news',
-                results
-              })
-            } else {
-              res.send({
-                success: false,
-                message: 'Fail to add news'
-              })
-            }
-          }
+        const { value, error } = schema.validate(req.body)
+        const { title, news, category } = value
+        if (error) {
+          res.send({
+            success: false,
+            message: `${error}`
+          })
         } else {
-          const image = `uploads/${req.file.filename}` // ambil filename dari helpers yang sudah di edit
-          const checkNews = await News.findAll({ where: { title: title } })
-          if (checkNews.length > 0) {
-            res.send({
-              success: false,
-              message: 'News alreadey exists'
-            })
-          } else {
-            const data = {
-              title, news, category, image: image
-            }
-            const results = await News.create(data)
-            if (results) {
-              res.send({
-                success: true,
-                message: 'Success add news',
-                results
-              })
-            } else {
+          if (req.file === undefined) {
+            const checkNews = await News.findAll({ where: { title: title } })
+            if (checkNews.length > 0) {
               res.send({
                 success: false,
-                message: 'Fail to add news'
+                message: 'News alreadey exists'
               })
+            } else {
+              const data = {
+                title, news, category, image: ''
+              }
+              const results = await News.create(data)
+              if (results) {
+                res.send({
+                  success: true,
+                  message: 'Success add news',
+                  results
+                })
+              } else {
+                res.send({
+                  success: false,
+                  message: 'Fail to add news'
+                })
+              }
+            }
+          } else {
+            const image = `uploads/${req.file.filename}` // ambil filename dari helpers yang sudah di edit
+            const checkNews = await News.findAll({ where: { title: title } })
+            if (checkNews.length > 0) {
+              res.send({
+                success: false,
+                message: 'News alreadey exists'
+              })
+            } else {
+              const data = {
+                title, news, category, image: image
+              }
+              const results = await News.create(data)
+              if (results) {
+                res.send({
+                  success: true,
+                  message: 'Success add news',
+                  results
+                })
+              } else {
+                res.send({
+                  success: false,
+                  message: 'Fail to add news'
+                })
+              }
             }
           }
         }
+      } else {
+        res.send({
+          success: false,
+          message: 'You are not a admin or jurnalis'
+        })
       }
     } catch (err) {
       res.send({
@@ -125,58 +132,123 @@ module.exports = {
 
   updateNews: async (req, res) => {
     try {
-      const { id } = req.params
-
-      const checkNews = await News.findAll({ where: { id: id } })
-      if (checkNews.length) {
-        const schema = joi.object({
-          title: joi.string().required(),
-          news: joi.string().required(),
-          category: joi.string().required()
-        })
-        const { value, error } = schema.validate(req.body)
-        const { title, news, category } = value
-        if (error) {
-          res.send({
-            success: false,
-            message: `${error}`
-          })
-        } else {
-          const getTitileFromDb = checkNews[0].title
-          if (title === getTitileFromDb || title !== getTitileFromDb) {
-            const findAnotherTitle = await News.findAll({ where: { title: { [Op.ne]: getTitileFromDb } } })
-            const mapsTitleAnother = findAnotherTitle.map(o => {
-              return o.title
+      const { role } = req.user.jwtToken
+      if (role === 'admin' || role === 'jurnalis') {
+        const { id } = req.params
+        const checkNews = await News.findAll({ where: { id: id } })
+        if (checkNews.length) {
+          if (req.file === undefined) {
+            const schema = joi.object({
+              title: joi.string().required(),
+              news: joi.string().required(),
+              category: joi.string().required()
             })
-            console.log('maps', mapsTitleAnother)
-            const someNews = await mapsTitleAnother.some(item => item === title)
-            if (someNews === true) {
+            const { value, error } = schema.validate(req.body)
+            const { title, news, category } = value
+            if (error) {
               res.send({
                 success: false,
-                message: 'News already exists'
+                message: `${error}`
               })
             } else {
-              const data = {
-                title, news, category
-              }
-              const updateNews = await News.update(data, { where: { id: id } })
-              if (updateNews.length > 0) {
-                const getNewsAfterUpdate = await News.findAll({ where: { id: id } })
-                if (getNewsAfterUpdate.length) {
+              const getTitileFromDb = checkNews[0].title
+              if (title === getTitileFromDb || title !== getTitileFromDb) {
+                const findAnotherTitle = await News.findAll({ where: { title: { [Op.ne]: getTitileFromDb } } })
+                const mapsTitleAnother = findAnotherTitle.map(o => {
+                  return o.title
+                })
+                console.log('maps', mapsTitleAnother)
+                const someNews = await mapsTitleAnother.some(item => item === title)
+                if (someNews === true) {
                   res.send({
-                    success: true,
-                    message: 'Updated succrssfully',
-                    results: getNewsAfterUpdate[0]
+                    success: false,
+                    message: 'News already exists'
                   })
+                } else {
+                  const data = {
+                    title, news, category
+                  }
+                  const updateNews = await News.update(data, { where: { id: id } })
+                  if (updateNews.length > 0) {
+                    const getNewsAfterUpdate = await News.findAll({ where: { id: id } })
+                    if (getNewsAfterUpdate.length) {
+                      res.send({
+                        success: true,
+                        message: 'Updated succrssfully',
+                        results: getNewsAfterUpdate[0]
+                      })
+                    } else {
+                      res.send({
+                        success: false,
+                        message: 'Fail to update news'
+                      })
+                    }
+                  }
+                }
+              }
+            }
+          } else {
+            const image = `uploads/${req.file.filename}`
+            const schema = joi.object({
+              title: joi.string().required(),
+              news: joi.string().required(),
+              category: joi.string().required()
+            })
+            const { value, error } = schema.validate(req.body)
+            const { title, news, category } = value
+            if (error) {
+              res.send({
+                success: false,
+                message: `${error}`
+              })
+            } else {
+              const getTitileFromDb = checkNews[0].title
+              if (title === getTitileFromDb || title !== getTitileFromDb) {
+                const findAnotherTitle = await News.findAll({ where: { title: { [Op.ne]: getTitileFromDb } } })
+                const mapsTitleAnother = findAnotherTitle.map(o => {
+                  return o.title
+                })
+                //   console.log('maps', mapsTitleAnother)
+                const someNews = await mapsTitleAnother.some(item => item === title)
+                if (someNews === true) {
+                  res.send({
+                    success: false,
+                    message: 'News already exists'
+                  })
+                } else {
+                  const data = {
+                    title, news, category, image: image
+                  }
+                  const updateNews = await News.update(data, { where: { id: id } })
+                  if (updateNews.length > 0) {
+                    const getNewsAfterUpdate = await News.findAll({ where: { id: id } })
+                    if (getNewsAfterUpdate.length) {
+                      res.send({
+                        success: true,
+                        message: 'Updated succrssfully',
+                        results: getNewsAfterUpdate[0]
+                      })
+                    }
+                  } else {
+                    res.send({
+                      success: false,
+                      message: 'Fail to update news'
+                    })
+                  }
                 }
               }
             }
           }
+        } else {
+          res.send({
+            success: false,
+            message: 'Data not found'
+          })
         }
       } else {
         res.send({
           success: false,
-          message: 'Data not found'
+          message: 'You are not admin'
         })
       }
     } catch (err) {
@@ -189,25 +261,33 @@ module.exports = {
 
   deleteNews: async (req, res) => {
     try {
-      const { id } = req.params
-      const checkNews = await News.findAll({ where: { id: id } })
-      if (checkNews.length > 0) {
-        const deleteNews = await News.destroy({ where: { id: id } })
-        if (deleteNews) {
-          res.send({
-            success: true,
-            message: 'Delete successfully'
-          })
+      const { role } = req.user.jwtToken
+      if (role === 'admin' || role === 'jurnalis') {
+        const { id } = req.params
+        const checkNews = await News.findAll({ where: { id: id } })
+        if (checkNews.length > 0) {
+          const deleteNews = await News.destroy({ where: { id: id } })
+          if (deleteNews) {
+            res.send({
+              success: true,
+              message: 'Delete successfully'
+            })
+          } else {
+            res.send({
+              success: false,
+              message: 'Fail to delete news'
+            })
+          }
         } else {
           res.send({
             success: false,
-            message: 'Fail to delete news'
+            message: 'News not found'
           })
         }
       } else {
         res.send({
           success: false,
-          message: 'News not found'
+          message: 'You are not admin'
         })
       }
     } catch (err) {
